Main/AGENT_MAIN/requirements.txt
azure-functions
openai
azure-data-tables